requirements.txt
customtkinter==5.2.2
pandas==2.1.4
openpyxl==3.1.2
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pydrive2==1.19.0
Pillow==10.1.0